One step away from completion...I think
Managed to get everything how I want it near enough. Not completely satisfied with the scoreBox.innerHTML during some screen sizes but not the end of the world. The biggest problem I have now is the soundOn function. Can't get the damn thing to turn the sound back on. I'll figure it out eventually.