So close yet so far away!
Managed to overcome the monumental task of turning the audio on and off with just one button. However, now the setTimeout doesn't seem to be triggering the music upon page load. 

I've even managed to sort the responsive stuff out; the project looks fine at a width of just 305px. 

Was hoping to have this completely done by today, but I'll have to wait until Monday to speak with my mentor. Royal pain in the backside!